+++ CHANGELOG.md
@@ -4,12 +4,23 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 ### Added
-- ability to set custom output manipulations
-
+- more html
+- css for styling
 ### Changed
+### Removed
+- memory limit hack
+
+
+## [0.1.1] - 2017-05-11
+### Added
+- iterating variable parser.
+### Changed
+- new parser now generates output instead of modifying var_dump() output.
+
 
 ## [0.1.0] - 2016-09-23
 ### Added
 - First Release... beta, I guess?
 
-[unreleased]: https://github.com/radsectors/urp/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/radsectors/urp/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/radsectors/urp/compare/v0.1.0...v0.2.0